+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # ctw-component-library
 
+## 0.48.4
+
+### Patch Changes
+
+- cb98edb6: Export patientSearch component.
+
+## 0.48.3
+
+### Patch Changes
+
+- d9191657: Add patient search component.
+
 ## 0.48.2
 
 ### Patch Changes